@@ -110,44 +110,44 @@
         "zone.js": "~0.11.5"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~13.3.0",
-        "@angular-eslint/builder": "^13.1.0",
-        "@angular-eslint/eslint-plugin": "^13.1.0",
-        "@angular-eslint/eslint-plugin-template": "^13.1.0",
-        "@angular-eslint/schematics": "^13.1.0",
-        "@angular-eslint/template-parser": "^13.1.0",
-        "@angular/cli": "~13.3.0",
-        "@angular/compiler-cli": "~13.3.0",
-        "@commitlint/cli": "^16.2.3",
-        "@commitlint/config-angular": "^16.2.3",
+        "@angular-devkit/build-angular": "~13.3.6",
+        "@angular-eslint/builder": "^13.2.1",
+        "@angular-eslint/eslint-plugin": "^13.2.1",
+        "@angular-eslint/eslint-plugin-template": "^13.2.1",
+        "@angular-eslint/schematics": "^13.2.1",
+        "@angular-eslint/template-parser": "^13.2.1",
+        "@angular/cli": "~13.3.6",
+        "@angular/compiler-cli": "~13.3.9",
+        "@commitlint/cli": "^17.0.0",
+        "@commitlint/config-angular": "^17.0.0",
         "@compodoc/compodoc": "^1.1.19",
-        "@types/d3": "^7.1.0",
-        "@types/jasmine": "~4.0.0",
+        "@types/d3": "^7.4.0",
+        "@types/jasmine": "~4.0.3",
         "@types/node": "^15.14.9",
-        "@typescript-eslint/eslint-plugin": "^5.16.0",
-        "@typescript-eslint/parser": "^5.16.0",
+        "@typescript-eslint/eslint-plugin": "^5.26.0",
+        "@typescript-eslint/parser": "^5.26.0",
         "angular-cli-ghpages": "^1.0.0",
         "conventional-recommended-bump": "^6.1.0",
         "cross-var": "^1.1.0",
-        "eslint": "^8.11.0",
+        "eslint": "^8.16.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-angular": "^4.1.0",
-        "eslint-plugin-import": "^2.25.4",
-        "eslint-plugin-jsdoc": "^38.0.6",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-jsdoc": "^39.3.0",
         "eslint-plugin-prettier": "^4.0.0",
-        "gzipper": "^7.0.0",
-        "husky": "^7.0.4",
-        "jasmine-core": "~4.0.1",
-        "karma": "~6.3.17",
+        "gzipper": "^7.1.0",
+        "husky": "^8.0.1",
+        "jasmine-core": "~4.1.1",
+        "karma": "~6.3.20",
         "karma-chrome-launcher": "~3.1.1",
         "karma-coverage": "^2.2.0",
-        "karma-jasmine": "~4.0.1",
+        "karma-jasmine": "~5.0.1",
         "karma-jasmine-html-reporter": "^1.7.0",
-        "lint-staged": "^12.3.7",
-        "prettier": "^2.6.0",
+        "lint-staged": "^12.4.1",
+        "prettier": "^2.6.2",
         "source-map-explorer": "^2.5.2",
-        "standard-version": "^9.3.2",
-        "typescript": "~4.6.2",
+        "standard-version": "^9.5.0",
+        "typescript": "~4.6.4",
         "webpack-bundle-analyzer": "^4.5.0"
     },
     "engines": {
