@@ -101,8 +101,10 @@ describe('ApiService', () => {
         it('... should issue a mocked http get request', waitForAsync(() => {
             const testData: Data = { name: 'TestData' };
 
-            httpClient.get<Data>('/foo/bar').subscribe(data => {
-                expect(data).toEqual(testData);
+            httpClient.get<Data>('/foo/bar').subscribe({
+                next: data => {
+                    expect(data).toEqual(testData);
+                },
             });
 
             // Match the request url
@@ -192,15 +194,15 @@ describe('ApiService', () => {
                     const expectedApiServiceResult = new ApiServiceResult();
 
                     // Call service function (success)
-                    apiService
-                        .getApiResponse(ApiServiceResult, queryPath, queryHttpParams)
-                        .subscribe((res: ApiServiceResult) => {
+                    apiService.getApiResponse(ApiServiceResult, queryPath, queryHttpParams).subscribe({
+                        next: (res: ApiServiceResult) => {
                             expect(res).toBeTruthy();
                             expect(res.status).toEqual(expectedApiServiceResult.status);
                             expect(res.statusText).toEqual(expectedApiServiceResult.statusText);
                             expect(res.url).toEqual(expectedApiServiceResult.url);
                             expect(res.body).toBeUndefined();
-                        });
+                        },
+                    });
 
                     // Expect one request to url with given settings
                     const call = httpTestingController.expectOne(
@@ -216,12 +218,12 @@ describe('ApiService', () => {
                     expectedJsonResponse.lang = 'en';
 
                     // Call service function (success)
-                    apiService
-                        .getApiResponse(UserDataJson, queryPath, queryHttpParams)
-                        .subscribe((res: UserDataJson) => {
+                    apiService.getApiResponse(UserDataJson, queryPath, queryHttpParams).subscribe({
+                        next: (res: UserDataJson) => {
                             expect(res).toBeTruthy();
                             expect(res).toEqual(expectedJsonResponse);
-                        });
+                        },
+                    });
 
                     // Expect one request to url with given settings
                     const call = httpTestingController.expectOne(
@@ -235,7 +237,7 @@ describe('ApiService', () => {
 
             describe('fail', () => {
                 // See http://www.syntaxsuccess.com/viewarticle/unit-testing-rxjs-retries
-                it('... should throw default ApiServiceError if httpGet fails with undefined handler', done => {
+                it('... should throw default ApiServiceError if httpGet fails with undefined handler', waitForAsync(() => {
                     const expectedErrorMsg = 'should fail HTTP response';
                     expectedApiServiceError = new ApiServiceError();
 
@@ -243,21 +245,21 @@ describe('ApiService', () => {
                     spyOn(httpClient, 'get').and.returnValue(observableThrowError(() => expectedApiServiceError));
 
                     // Call service function (fail)
-                    apiService.getApiResponse(UserDataJson, queryPath).subscribe(
-                        res => fail(expectedErrorMsg),
-                        (error: ApiServiceError) => {
-                            expectErrorResponse(error, expectedApiServiceError);
-                            done();
-                        }
-                    );
+                    apiService.getApiResponse(UserDataJson, queryPath).subscribe({
+                        next: () => fail(expectedErrorMsg),
+                        error: (err: ApiServiceError) => {
+                            expectErrorResponse(err, expectedApiServiceError);
+                        },
+                        complete: () => fail(expectedErrorMsg),
+                    });
 
                     // Expect no request to url with given settings
                     httpTestingController.expectNone(
                         (req: HttpRequest<any>) => req.method === 'GET' && req.url === expectedUrl
                     );
-                });
+                }));
 
-                it('... should throw specified ApiServiceError if httpGet fails with specified handler', done => {
+                it('... should throw specified ApiServiceError if httpGet fails with specified handler', waitForAsync(() => {
                     const expectedErrorMsg = 'should fail HTTP response with 500 error';
                     expectedApiServiceError = createApiServiceError(500, 'Internal Server Error', true);
 
@@ -267,32 +269,33 @@ describe('ApiService', () => {
                     );
 
                     // Call service function (fail)
-                    apiService.getApiResponse(UserDataJson, queryPath).subscribe(
-                        res => fail(expectedErrorMsg),
-                        (error: ApiServiceError) => {
+                    apiService.getApiResponse(UserDataJson, queryPath).subscribe({
+                        next: () => fail(expectedErrorMsg),
+                        error: (err: ApiServiceError) => {
                             expect(apiService.getApiResponse).toHaveBeenCalled();
-                            expectErrorResponse(error, expectedApiServiceError);
-                            done();
-                        }
-                    );
+                            expectErrorResponse(err, expectedApiServiceError);
+                        },
+                        complete: () => fail(expectedErrorMsg),
+                    });
 
                     // Expect no request to url with given settings
                     httpTestingController.expectNone(
                         (req: HttpRequest<any>) => req.method === 'GET' && req.url === expectedUrl
                     );
-                });
+                }));
 
                 it("... should return 'ApiServiceError' for 401 Unauthorized", waitForAsync(() => {
                     const expectedErrorMsg = 'should fail HTTP response with 401 error';
                     expectedApiServiceError = createApiServiceError(401, 'Unauthorized');
 
                     // Call service function (fail)
-                    apiService.getApiResponse(UserDataJson, queryPath, queryHttpParams).subscribe(
-                        res => fail(expectedErrorMsg),
-                        (error: ApiServiceError) => {
-                            expectErrorResponse(error, expectedApiServiceError);
-                        }
-                    );
+                    apiService.getApiResponse(UserDataJson, queryPath, queryHttpParams).subscribe({
+                        next: () => fail(expectedErrorMsg),
+                        error: (err: ApiServiceError) => {
+                            expectErrorResponse(err, expectedApiServiceError);
+                        },
+                        complete: () => fail(expectedErrorMsg),
+                    });
 
                     // Expect one request to url with given settings
                     const call = httpTestingController.expectOne(
@@ -308,12 +311,13 @@ describe('ApiService', () => {
                     expectedApiServiceError = createApiServiceError(404, 'Not found');
 
                     // Call service function (fail)
-                    apiService.getApiResponse(UserDataJson, queryPath, queryHttpParams).subscribe(
-                        res => fail(expectedErrorMsg),
-                        (error: ApiServiceError) => {
-                            expectErrorResponse(error, expectedApiServiceError);
-                        }
-                    );
+                    apiService.getApiResponse(UserDataJson, queryPath, queryHttpParams).subscribe({
+                        next: () => fail(expectedErrorMsg),
+                        error: (err: ApiServiceError) => {
+                            expectErrorResponse(err, expectedApiServiceError);
+                        },
+                        complete: () => fail(expectedErrorMsg),
+                    });
 
                     // Expect one request to url with given settings
                     const call = httpTestingController.expectOne(
