@@ -252,57 +252,61 @@ describe('EditionViewComponent (DONE)', () => {
                 expectSpyCall(getSelectionsFromRouteSpy, 1);
             });
 
-            it('... should get isRowTableView$ (via EditionService)', done => {
+            it('... should get isRowTableView$ (via EditionService)', waitForAsync(() => {
                 expectSpyCall(getSelectionsFromRouteSpy, 1);
                 expectSpyCall(editionServiceGetIsRowTableViewSpy, 1);
 
                 expect(component.isRowTableView$).toBeDefined();
-                component.isRowTableView$.subscribe((isView: boolean) => {
-                    expect(isView)
-                        .withContext(`should equal ${expectedIsRowTableView}`)
-                        .toEqual(expectedIsRowTableView);
-                    done();
+                component.isRowTableView$.subscribe({
+                    next: (isView: boolean) => {
+                        expect(isView)
+                            .withContext(`should equal ${expectedIsRowTableView}`)
+                            .toEqual(expectedIsRowTableView);
+                    },
                 });
-            });
+            }));
 
-            it('... should get selectedEditionSeries$ (via EditionService)', done => {
+            it('... should get selectedEditionSeries$ (via EditionService)', waitForAsync(() => {
                 expectSpyCall(getSelectionsFromRouteSpy, 1);
                 expectSpyCall(editionServiceGetSelectedEditionSeriesSpy, 1);
 
                 expect(component.selectedEditionSeries$).toBeDefined();
-                component.selectedEditionSeries$.subscribe((series: EditionSeriesRoutes) => {
-                    expect(series)
-                        .withContext(`should equal ${expectedSelectedEditionSeries}`)
-                        .toEqual(expectedSelectedEditionSeries);
-                    done();
+                component.selectedEditionSeries$.subscribe({
+                    next: (series: EditionSeriesRoutes) => {
+                        expect(series)
+                            .withContext(`should equal ${expectedSelectedEditionSeries}`)
+                            .toEqual(expectedSelectedEditionSeries);
+                    },
                 });
-            });
+            }));
 
-            it('... should get selectedEditionSection$ (via EditionService)', done => {
+            it('... should get selectedEditionSection$ (via EditionService)', waitForAsync(() => {
                 expectSpyCall(getSelectionsFromRouteSpy, 1);
                 expectSpyCall(editionServiceGetSelectedEditionSectionSpy, 1);
 
                 expect(component.selectedEditionSection$).toBeDefined();
-                component.selectedEditionSection$.subscribe((section: EditionRoute) => {
-                    expect(section)
-                        .withContext(`should equal ${expectedSelectedEditionSection}`)
-                        .toEqual(expectedSelectedEditionSection);
-                    done();
+                component.selectedEditionSection$.subscribe({
+                    next: (section: EditionRoute) => {
+                        expect(section)
+                            .withContext(`should equal ${expectedSelectedEditionSection}`)
+                            .toEqual(expectedSelectedEditionSection);
+                    },
                 });
-            });
+            }));
 
-            it('... should get selectedEditionComplex$ (via EditionService)', done => {
+            it('... should get selectedEditionComplex$ (via EditionService)', waitForAsync(() => {
                 expectSpyCall(getSelectionsFromRouteSpy, 1);
                 expectSpyCall(editionServiceGetEditionComplexSpy, 1);
 
                 expect(component.selectedEditionComplex$).toBeDefined();
-                component.selectedEditionComplex$.subscribe((work: EditionWork) => {
-                    expect(work)
-                        .withContext(`should equal ${EditionWorks[expectedSelectedEditionComplexId]}`)
-                        .toEqual(EditionWorks[expectedSelectedEditionComplexId]);
-                    done();
+                component.selectedEditionComplex$.subscribe({
+                    next: (work: EditionWork) => {
+                        expect(work)
+                            .withContext(`should equal ${EditionWorks[expectedSelectedEditionComplexId]}`)
+                            .toEqual(EditionWorks[expectedSelectedEditionComplexId]);
+                    },
                 });
-            });
+            }));
         });
 
         describe('#routeToSideNav', () => {
